desagregacion de views en core y vistas dev
