+++ core/urls.py
@@ -1,16 +1,36 @@
 from django.urls import path
-from .views import health, db_health_check, hello_world, htmx_demo, login_view, logout_view, dashboard_view, index, dev_view
+
+# Importaciones directas desde módulos específicos
+from .views.home import index, dashboard_view
+from .views.auth import login_view, logout_view
+from .views.health import health, db_health_check
+from .views.demos import hello_world, htmx_demo
+from .views.dev import dev_view, dev_udns, dev_sectors, dev_categories, dev_hierarchy, dev_purchase_workflow
 
 app_name = 'core'
 
 urlpatterns = [
+    # Vistas principales (home.py)
     path('', index, name='index'),
-    path('health/', health, name='health'),
-    path('health/db/', db_health_check, name='db_health_check'),
-    path('hello/', hello_world, name='hello_world'),
-    path('htmx-demo/', htmx_demo, name='htmx_demo'),
+    path('dashboard/', dashboard_view, name='dashboard'),
+    
+    # Vistas de autenticación (auth.py)
     path('login/', login_view, name='login'),
     path('logout/', logout_view, name='logout'),
-    path('dashboard/', dashboard_view, name='dashboard'),
+    
+    # Health checks (health.py)
+    path('health/', health, name='health'),
+    path('health/db/', db_health_check, name='db_health_check'),
+    
+    # Demos y pruebas (demos.py)
+    path('hello/', hello_world, name='hello_world'),
+    path('htmx-demo/', htmx_demo, name='htmx_demo'),
+    
+    # Vistas de desarrollo (dev.py)
     path('dev/', dev_view, name='dev'),
+    path('dev/udns/', dev_udns, name='dev_udns'),
+    path('dev/sectors/', dev_sectors, name='dev_sectors'),
+    path('dev/categories/', dev_categories, name='dev_categories'),
+    path('dev/hierarchy/', dev_hierarchy, name='dev_hierarchy'),
+    path('dev/purchase-workflow/', dev_purchase_workflow, name='dev_purchase_workflow'),
 ]
