+++ templates/core/dev/sectors.html
@@ -1,0 +1,100 @@
+<section class="dev-section">
+    <h2 class="section-title">Sectores</h2>
+    <div class="sector-list">
+        <div class="sector-item">
+            <span class="tag tag-sector">Full</span>
+            <div class="sector-udns">
+                <span class="tag tag-udn-mini">Km 1151</span>
+                <span class="tag tag-udn-mini">Las Bóvedas</span>
+            </div>
+        </div>
+
+        <div class="sector-separator"></div>
+
+        <div class="sector-item">
+            <span class="tag tag-sector">Playa</span>
+            <div class="sector-udns">
+                <span class="tag tag-udn-mini">Km 1151</span>
+                <span class="tag tag-udn-mini">Las Bóvedas</span>
+            </div>
+        </div>
+
+        <div class="sector-separator"></div>
+
+        <div class="sector-item">
+            <span class="tag tag-sector">Administración</span>
+            <div class="sector-udns">
+                <span class="tag tag-udn-mini">Km 1151</span>
+                <span class="tag tag-udn-mini">Las Bóvedas</span>
+                <span class="tag tag-udn-mini">Oficina Espejo</span>
+            </div>
+        </div>
+
+        <div class="sector-separator"></div>
+
+        <div class="sector-item">
+            <span class="tag tag-sector">Parador</span>
+            <div class="sector-udns">
+                <span class="tag tag-udn-mini">Km 1151</span>
+            </div>
+        </div>
+    </div>
+</section>
+
+<style>
+    .dev-section {
+        @apply mb-12;
+    }
+
+    .section-title {
+        @apply text-xl font-medium mb-6;
+        color: var(--theme-text-secondary);
+        font-family: var(--font-ui);
+    }
+
+    .tag {
+        @apply text-xs font-medium py-1.5 px-3 rounded-full leading-none whitespace-nowrap border-2 transition-all duration-200;
+        font-family: var(--font-ui);
+    }
+
+    .tag-sector {
+        background-color: var(--theme-primary-light);
+        color: var(--theme-primary);
+        border-color: var(--theme-primary);
+    }
+
+    .tag-sector:hover {
+        background-color: var(--theme-primary);
+        color: var(--theme-surface);
+        @apply cursor-pointer;
+        transform: translateY(-1px);
+    }
+
+    .tag-udn-mini {
+        @apply py-1 px-2 rounded-full border;
+        background-color: #fef3c7;
+        color: #d97706;
+        border-color: #fcd34d;
+        font-family: var(--font-ui);
+        font-size: 10px;
+        font-weight: 500;
+    }
+
+    .sector-list {
+        @apply space-y-0;
+    }
+
+    .sector-item {
+        @apply flex flex-col gap-3 py-3;
+    }
+
+    .sector-udns {
+        @apply flex flex-wrap gap-2 pl-8 ml-4;
+    }
+
+    .sector-separator {
+        @apply h-px my-4;
+        background-color: var(--theme-border);
+        opacity: 0.5;
+    }
+</style>
