+++ templates/core/dev/udns.html
@@ -1,0 +1,38 @@
+<section class="dev-section">
+    <h2 class="section-title">UDNs</h2>
+    <div class="tag-container">
+        <span class="tag tag-udn">Km 1151</span>
+        <span class="tag tag-udn">Las Bóvedas</span>
+        <span class="tag tag-udn">Oficina Espejo</span>
+    </div>
+</section>
+
+<style>
+    .dev-section {
+        @apply mb-12;
+    }
+
+    .section-title {
+        @apply text-xl font-medium mb-6;
+        color: var(--theme-text-secondary);
+        font-family: var(--font-ui);
+    }
+
+    .tag-container {
+        @apply flex flex-wrap gap-3;
+    }
+
+    .tag {
+        @apply text-xs font-medium py-1.5 px-3 rounded-full leading-none whitespace-nowrap border-2 transition-all duration-200;
+        font-family: var(--font-ui);
+    }
+
+    .tag-udn {
+        @apply bg-yellow-100 text-amber-600 border-amber-300;
+    }
+
+    .tag-udn:hover {
+        @apply bg-yellow-200 border-amber-400 cursor-pointer;
+        transform: translateY(-1px);
+    }
+</style>
